Move TablePagination to seperate component
