+++ ccm_web/client/src/components/ValidationErrorTable.tsx
@@ -1,15 +1,7 @@
 import React, { useEffect, useState } from 'react'
-import { createStyles, IconButton, makeStyles, Paper, Table, TableBody, TableCell, TableContainer, TableFooter, TableHead, TablePagination, TableRow, Theme, useTheme, withStyles } from '@material-ui/core'
-import { KeyboardArrowRight, KeyboardArrowLeft, LastPage as LastPageIcon, FirstPage as FirstPageIcon } from '@material-ui/icons'
+import { createStyles, Paper, Table, TableBody, TableCell, TableContainer, TableFooter, TableHead, TablePagination, TableRow, Theme, withStyles } from '@material-ui/core'
 import { grey } from '@material-ui/core/colors'
-import PropTypes from 'prop-types'
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexShrink: 0,
-    marginLeft: theme.spacing(2.5)
-  }
-}))
+import { TablePaginationActions } from './TablePagination'
 
 const StyledTableCell = withStyles((theme: Theme) =>
   createStyles({
@@ -46,7 +38,6 @@ const columns: TableHeaderColumnInfoShouldUseMatUIType[] = [
 ]
 
 function ErrorTable (props: ErrorTableProps): JSX.Element {
-  const classes = useStyles()
   const [tableRows, setTableRows] = useState<ValidationError[]>([])
   const [page, setPage] = React.useState<number>(0)
   const [rowsPerPage, setRowsPerPage] = React.useState(5)
@@ -68,66 +59,6 @@ function ErrorTable (props: ErrorTableProps): JSX.Element {
     const handleChangeRowsPerPage = (event: { target: { value: string } }): void => {
       setRowsPerPage(parseInt(event.target.value, rowsPerPage))
       setPage(0)
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    function TablePaginationActions (props: { count: any, page: any, rowsPerPage: any, onChangePage: any }): JSX.Element {
-      // const classes = useStyles()
-      const theme = useTheme()
-      const { count, page, rowsPerPage, onChangePage } = props
-
-      const handleFirstPageButtonClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>): void => {
-        onChangePage(event, 0)
-      }
-
-      const handleBackButtonClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>): void => {
-        onChangePage(event, page - 1)
-      }
-
-      const handleNextButtonClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>): void => {
-        // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
-        onChangePage(event, props.page + 1)
-      }
-
-      const handleLastPageButtonClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>): void => {
-        onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1))
-      }
-
-      return (
-        <div className={classes.root}>
-          <IconButton
-            onClick={handleFirstPageButtonClick}
-            disabled={page === 0}
-            aria-label="first page"
-          >
-            {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-          </IconButton>
-          <IconButton onClick={handleBackButtonClick} disabled={page === 0} aria-label="previous page">
-            {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-          </IconButton>
-          <IconButton
-            onClick={handleNextButtonClick}
-            disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-            aria-label="next page"
-          >
-            {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-          </IconButton>
-          <IconButton
-            onClick={handleLastPageButtonClick}
-            disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-            aria-label="last page"
-          >
-            {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-          </IconButton>
-        </div>
-      )
-    }
-
-    TablePaginationActions.propTypes = {
-      count: PropTypes.number.isRequired,
-      onChangePage: PropTypes.func.isRequired,
-      page: PropTypes.number.isRequired,
-      rowsPerPage: PropTypes.number.isRequired
     }
 
     return (
